AuthorizationStatus/ Screenshot Handling
Implemented authorizationStatus popup needed for album creation. Created handler functions to locate the screenshots, process them, and copy them to the album based on the key value in the formData.

TODO:
Unit Testing for authorizationStatus/ probably enhance the code better for full integration testing.

Unit Testing for ScreenshotHandler and physical device testing to properly ensure that the code works as simulator could not be as reliable.

LIMITATIONS: cannot move image to album due to API limitations.